backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
aiohttp>=3.9.0
aiosqlite==0.20.0
fastapi==0.111.0
geojson>=3.1.0
geopandas==0.14.4
huggingface_hub>=0.31.0
httpx==0.27.0
numpy==1.26.4
Pillow>=10.0.0
pydantic==2.7.1
python-multipart>=0.0.9
python-dotenv==1.0.1
requests==2.31.0
shapely==2.0.4
torch==2.3.1+cpu
torchvision>=0.18.1
uvicorn[standard]==0.29.0
websockets==12.0
